fix(neb): update engine headers

+++ neb/inc/com/centreon/engine/logging/debug_file.hh
@@ -34,7 +34,7 @@ namespace logging {
  */
 class debug_file : public com::centreon::logging::file {
  public:
-  debug_file(std::string const& path, long long max_size = 0);
+  debug_file(std::string const& path, uint64_t max_size = 0);
   ~debug_file() throw() override;
 
  private:
